[Security Solution] Fix rules table filtering by tags with special characters (#163767)
**Fixes: https://github.com/elastic/kibana/issues/163692**

## Summary

This PR fixes rules table filtering for tags with special characters like colon `category:tag`.

We didn't have this bug in `8.9` it was introduced by https://github.com/elastic/kibana/pull/160480 in `8.10` scope.

## Details

Coverage Overview Dashboard required very similar rule filtering functionality as rule table has so https://github.com/elastic/kibana/pull/160480 moved the common logic to `common` folder and started using `escapeKuery` from `@kbn/es-query` instead of custom `escapeKuery` defined in Security Solution in `x-pack/plugins/security_solution/public/common/lib/kuery/index.ts`. The difference is the first function escapes all the special characters `\():<>"*` but the second escapes only quotes but the value must be enclosed in quotes to be passed on. Both ways are correct according to [docs](https://www.elastic.co/guide/en/kibana/current/kuery-query.html). So escaping all the special characters and enclosing the value in quotes breaks rules search.

### How was it fixed?

Escaping related code in `x-pack/plugins/security_solution/common/utils/kql.ts` (`convertRulesFilterToKQL` in particular) was moved to `x-pack/plugins/security_solution/common/detection_engine/rule_management/rule_filtering.ts` as Coverage Overview Dashboard API endpoint and rules table UI share the same KQL helpers and refactored along the way to be much more transparent.

### Checklist

- [x] [Unit or functional tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html) were updated or added to match the most common scenarios

(cherry picked from commit 92ff716b750c2a47bf55ff280e0621e553f1f019)

# Conflicts:
#	x-pack/plugins/security_solution/public/detection_engine/rule_management/api/api.test.ts
