+++ x-pack/plugins/security_solution/common/detection_engine/rule_management/rule_fields.ts
