+++ x-pack/plugins/security_solution/common/utils/kql.ts
@@ -5,111 +5,51 @@
  * 2.0.
  */
 
-import { escapeKuery } from '@kbn/es-query';
-import type { Type } from '@kbn/securitysolution-io-ts-alerting-types';
-import { RuleExecutionStatus } from '../api/detection_engine';
-import {
-  ENABLED_FIELD,
-  LAST_RUN_OUTCOME_FIELD,
-  PARAMS_IMMUTABLE_FIELD,
-  PARAMS_TYPE_FIELD,
-  RULE_NAME_FIELD,
-  RULE_PARAMS_FIELDS,
-  TAGS_FIELD,
-} from '../rule_fields';
+import { flow, isString } from 'lodash/fp';
 
-export const KQL_FILTER_IMMUTABLE_RULES = `${PARAMS_IMMUTABLE_FIELD}: true`;
-export const KQL_FILTER_MUTABLE_RULES = `${PARAMS_IMMUTABLE_FIELD}: false`;
-export const KQL_FILTER_ENABLED_RULES = `${ENABLED_FIELD}: true`;
-export const KQL_FILTER_DISABLED_RULES = `${ENABLED_FIELD}: false`;
-
-interface RulesFilterOptions {
-  filter: string;
-  showCustomRules: boolean;
-  showElasticRules: boolean;
-  enabled: boolean;
-  tags: string[];
-  excludeRuleTypes: Type[];
-  ruleExecutionStatus: RuleExecutionStatus;
+/**
+ * Preparing an arbitrary KQL query param by quoting and escaping string values, stringifying non string values.
+ *
+ * See https://www.elastic.co/guide/en/kibana/current/kuery-query.html
+ *
+ * @param value
+ * @returns
+ */
+export function prepareKQLParam(value: string | number | boolean): string {
+  return isString(value) ? prepareKQLStringParam(value) : `${value}`;
 }
 
 /**
- * Convert rules filter options object to KQL query
+ * Prepares a string KQL query param by wrapping the value in quotes and making sure
+ * the quotes, tabs and new line symbols inside are escaped.
  *
- * @param filterOptions desired filters (e.g. filter/sortField/sortOrder)
+ * See https://www.elastic.co/guide/en/kibana/current/kuery-query.html
  *
- * @returns KQL string
+ * @param value a string param value intended to be passed to KQL
+ * @returns a quoted and escaped string param value
  */
-export function convertRulesFilterToKQL({
-  filter: searchTerm,
-  showCustomRules,
-  showElasticRules,
-  enabled,
-  tags,
-  excludeRuleTypes = [],
-  ruleExecutionStatus,
-}: Partial<RulesFilterOptions>): string {
-  const kql: string[] = [];
-
-  if (searchTerm?.length) {
-    kql.push(`(${convertRuleSearchTermToKQL(searchTerm)})`);
-  }
-
-  if (showCustomRules && showElasticRules) {
-    // if both showCustomRules && showElasticRules selected we omit filter, as it includes all existing rules
-  } else if (showElasticRules) {
-    kql.push(KQL_FILTER_IMMUTABLE_RULES);
-  } else if (showCustomRules) {
-    kql.push(KQL_FILTER_MUTABLE_RULES);
-  }
-
-  if (enabled !== undefined) {
-    kql.push(enabled ? KQL_FILTER_ENABLED_RULES : KQL_FILTER_DISABLED_RULES);
-  }
-
-  if (tags?.length) {
-    kql.push(convertRuleTagsToKQL(tags));
-  }
-
-  if (excludeRuleTypes.length) {
-    kql.push(`NOT ${convertRuleTypesToKQL(excludeRuleTypes)}`);
-  }
-
-  if (ruleExecutionStatus === RuleExecutionStatus.succeeded) {
-    kql.push(`${LAST_RUN_OUTCOME_FIELD}: "succeeded"`);
-  } else if (ruleExecutionStatus === RuleExecutionStatus['partial failure']) {
-    kql.push(`${LAST_RUN_OUTCOME_FIELD}: "warning"`);
-  } else if (ruleExecutionStatus === RuleExecutionStatus.failed) {
-    kql.push(`${LAST_RUN_OUTCOME_FIELD}: "failed"`);
-  }
-
-  return kql.join(' AND ');
+export function prepareKQLStringParam(value: string): string {
+  return `"${escapeKQLStringParam(value)}"`;
 }
 
-const SEARCHABLE_RULE_ATTRIBUTES = [
-  RULE_NAME_FIELD,
-  RULE_PARAMS_FIELDS.INDEX,
-  RULE_PARAMS_FIELDS.TACTIC_ID,
-  RULE_PARAMS_FIELDS.TACTIC_NAME,
-  RULE_PARAMS_FIELDS.TECHNIQUE_ID,
-  RULE_PARAMS_FIELDS.TECHNIQUE_NAME,
-  RULE_PARAMS_FIELDS.SUBTECHNIQUE_ID,
-  RULE_PARAMS_FIELDS.SUBTECHNIQUE_NAME,
-];
-
-export function convertRuleSearchTermToKQL(
-  searchTerm: string,
-  attributes = SEARCHABLE_RULE_ATTRIBUTES
-): string {
-  return attributes.map((param) => `${param}: "${escapeKuery(searchTerm)}"`).join(' OR ');
+/**
+ * Escapes string param intended to be passed to KQL. As official docs
+ * [here](https://www.elastic.co/guide/en/kibana/current/kuery-query.html) say
+ * `Certain characters must be escaped by a backslash (unless surrounded by quotes).` and
+ * `You must escape following characters: \():<>"*`.
+ *
+ * This function assumes the value is surrounded by quotes so it escapes quotes, tabs and new line symbols.
+ *
+ * @param param a string param value intended to be passed to KQL
+ * @returns an escaped string param value
+ */
+export function escapeKQLStringParam(value = ''): string {
+  return escapeStringValue(value);
 }
 
-export function convertRuleTagsToKQL(tags: string[]): string {
-  return `${TAGS_FIELD}:(${tags.map((tag) => `"${escapeKuery(tag)}"`).join(' AND ')})`;
-}
+const escapeQuotes = (val: string) => val.replace(/["]/g, '\\$&'); // $& means the whole matched string
 
-export function convertRuleTypesToKQL(ruleTypes: Type[]): string {
-  return `${PARAMS_TYPE_FIELD}: (${ruleTypes
-    .map((ruleType) => `"${escapeKuery(ruleType)}"`)
-    .join(' OR ')})`;
-}
+const escapeTabs = (val: string) =>
+  val.replace(/\t/g, '\\t').replace(/\r/g, '\\r').replace(/\n/g, '\\n');
+
+const escapeStringValue = flow(escapeQuotes, escapeTabs);
