+++ x-pack/plugins/security_solution/public/common/lib/kuery/index.test.ts
@@ -6,64 +6,8 @@
  */
 
 import expect from '@kbn/expect';
-import { convertToBuildEsQuery, escapeKuery } from '.';
+import { convertToBuildEsQuery } from '.';
 import { mockIndexPattern } from '../../mock';
-
-describe('Kuery escape', () => {
-  it('should not remove white spaces quotes', () => {
-    const value = ' netcat';
-    const expected = ' netcat';
-    expect(escapeKuery(value)).to.be(expected);
-  });
-
-  it('should escape quotes', () => {
-    const value = 'I said, "Hello."';
-    const expected = 'I said, \\"Hello.\\"';
-    expect(escapeKuery(value)).to.be(expected);
-  });
-
-  it('should escape special characters', () => {
-    const value = `This \\ has (a lot of) <special> characters, don't you *think*? "Yes."`;
-    const expected = `This \\ has (a lot of) <special> characters, don't you *think*? \\"Yes.\\"`;
-    expect(escapeKuery(value)).to.be(expected);
-  });
-
-  it('should NOT escape keywords', () => {
-    const value = 'foo and bar or baz not qux';
-    const expected = 'foo and bar or baz not qux';
-    expect(escapeKuery(value)).to.be(expected);
-  });
-
-  it('should NOT escape keywords next to each other', () => {
-    const value = 'foo and bar or not baz';
-    const expected = 'foo and bar or not baz';
-    expect(escapeKuery(value)).to.be(expected);
-  });
-
-  it('should not escape keywords without surrounding spaces', () => {
-    const value = 'And this has keywords, or does it not?';
-    const expected = 'And this has keywords, or does it not?';
-    expect(escapeKuery(value)).to.be(expected);
-  });
-
-  it('should NOT escape uppercase keywords', () => {
-    const value = 'foo AND bar';
-    const expected = 'foo AND bar';
-    expect(escapeKuery(value)).to.be(expected);
-  });
-
-  it('should escape special characters and NOT keywords', () => {
-    const value = 'Hello, "world", and <nice> to meet you!';
-    const expected = 'Hello, \\"world\\", and <nice> to meet you!';
-    expect(escapeKuery(value)).to.be(expected);
-  });
-
-  it('should escape newlines and tabs', () => {
-    const value = 'This\nhas\tnewlines\r\nwith\ttabs';
-    const expected = 'This\\nhas\\tnewlines\\r\\nwith\\ttabs';
-    expect(escapeKuery(value)).to.be(expected);
-  });
-});
 
 describe('convertToBuildEsQuery', () => {
   /**
