+++ x-pack/plugins/security_solution/public/explore/hosts/pages/details/helpers.ts
@@ -6,24 +6,6 @@
  */
 
 import type { Filter } from '@kbn/es-query';
-import { escapeQueryValue } from '../../../../common/lib/kuery';
-
-/** Returns the kqlQueryExpression for the `Events` widget on the `Host Details` page */
-export const getHostDetailsEventsKqlQueryExpression = ({
-  filterQueryExpression,
-  hostName,
-}: {
-  filterQueryExpression: string;
-  hostName: string;
-}): string => {
-  if (filterQueryExpression.length) {
-    return `${filterQueryExpression}${
-      hostName.length ? ` and host.name: ${escapeQueryValue(hostName)}` : ''
-    }`;
-  } else {
-    return hostName.length ? `host.name: ${escapeQueryValue(hostName)}` : '';
-  }
-};
 
 export const getHostDetailsPageFilters = (hostName: string): Filter[] => [
   {
