+++ x-pack/plugins/security_solution/public/timelines/components/timeline/helpers.tsx
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { isEmpty } from 'lodash/fp';
+import { isEmpty, isNumber } from 'lodash/fp';
 
 import {
   elementOrChildrenHasFocus,
@@ -15,11 +15,10 @@ import {
   stopPropagationAndPreventDefault,
 } from '@kbn/timelines-plugin/public';
 
+import { prepareKQLParam, prepareKQLStringParam } from '../../../../common/utils/kql';
 import { assertUnreachable } from '../../../../common/utility_types';
 import type { BrowserFields } from '../../../common/containers/source';
 import {
-  escapeQueryValue,
-  isNumber,
   convertDateFieldToQuery,
   checkIfFieldTypeIsDate,
   convertNestedFieldToQuery,
@@ -263,7 +262,7 @@ export const buildIsQueryMatch = ({
   } else if (checkIfFieldTypeIsDate(field, browserFields)) {
     return convertDateFieldToQuery(field, value);
   } else {
-    return `${field} : ${isNumber(value) ? value : escapeQueryValue(value)}`;
+    return `${field} : ${prepareKQLParam(value)}`;
   }
 };
 
@@ -291,7 +290,7 @@ export const buildIsOneOfQueryMatch = ({
   const trimmedField = field.trim();
   if (value.length) {
     return `${trimmedField} : (${value
-      .map((item) => (isNumber(item) ? Number(item) : `${escapeQueryValue(String(item).trim())}`))
+      .map((item) => (isNumber(item) ? item : prepareKQLStringParam(String(item).trim())))
       .join(' OR ')})`;
   }
   return `${trimmedField} : ''`;
